Another Restructurization: -Fixed project filters Lexic Analyzer: -Added solution project to include directory UnitTests: -Added Compiler project to include directory
Lexic Analyzer:
- Added GetToken()
