+++ Compiler/LexicAnalyzer/lists/operators.txt
@@ -1,0 +1,43 @@
+!
+$
+%
+^
+&
+*
+-
++
+=
+<
+<<
+<<=
+>
+>>
+>>=
+/
+~
+|
+@
+++
++=
+->
+->=
+--
+-=
+==
+<=
+>=
+**
+*=
+**=
+//
+/=
+//=
+^=
+&=
+&&
+|=
+||
+%=
+and
+or
+not
