+++ Compiler/LexicAnalyzer/lists/punctuations.txt
@@ -1,0 +1,9 @@
+,
+;
+:
+{
+}
+(
+)
+[
+]
