+++ Compiler/LexicAnalyzer/lists/reserved_ids.txt
@@ -1,0 +1,37 @@
+if
+else
+elif
+switch
+for
+while
+do
+return
+break
+goto
+continue
+try
+throw
+catch
+case
+default
+new
+delete
+import
+int8
+int16
+int32
+int64
+unsigned
+double
+float
+char
+let
+const
+var
+void
+and
+or
+not
+func
+NIL
+NULL
